requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.101.0
    # via -r requirements.in
paddleocr==2.7.0.0
    # via -r requirements.in
paddlepaddle==2.6.1
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
uvicorn==0.23.2
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
numpy==1.23.5
    # via -r requirements.in
python-docx==1.1.2
    # via -r requirements.in
opencv-python==4.6.0.66
    # via -r requirements.in
ultralytics==8.2.34
    # via -r requirements.in
onnx==1.16.1
    # via -r requirements.in
onnxruntime==1.18.0
    # via -r requirements.in
# The following packages are considered to be unsafe in a requirements file:
# setuptools
